+++ bin/tasks/replaces.sh
@@ -25,11 +25,11 @@ read -p "${BOLDYELLOW}Do we use comments in this project? (y/n)${TXTRESET} " yn
     echo ' '
   fi
 
-echo "${YELLOW}Running project gulp styles tasks once...${TXTRESET}"
-cd ${PROJECT_PATH}
+echo "${YELLOW}Running theme gulp styles tasks once...${TXTRESET}"
+cd ${PROJECT_THEME_PATH}
 gulp devstyles
 gulp prodstyles
 
-echo "${YELLOW}Running project gulp scripts task once...${TXTRESET}"
-cd ${PROJECT_PATH}
+echo "${YELLOW}Running theme gulp scripts task once...${TXTRESET}"
+cd ${PROJECT_THEME_PATH}
 gulp js